Web/requirements-dev.txt
appdirs==1.4.4
asgiref==3.2.10
attrs==19.3.0
autobahn==20.4.3
Automat==20.2.0
certifi==2020.6.20
cffi==1.14.0
constantly==15.1.0
cryptography==2.9.2
distlib==0.3.1
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.1
django-cors-headers==3.5.0
django-liveserver==0.1a2
django-pyodbc-azure==2.1.0.0
django-redis==4.11.0
djangorestframework==3.11.1
djangorestframework-simplejwt==4.4.0
filelock==3.0.12
html5lib==1.1
hyperlink==19.0.0
idna==2.9
incremental==17.5.0
numpy==1.19.0
Pillow==7.2.0
pipenv==2020.8.13
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
PyJWT==1.7.1
pyodbc==4.0.30
pyOpenSSL==19.1.0
PyPDF2==1.26.0
python-decouple==3.3
pytz==2020.1
redis==3.5.2
reportlab==3.5.44
service-identity==18.1.0
six==1.14.0
sqlparse==0.3.1
static3==0.7.0
txaio==20.4.1
virtualenv==20.0.30
virtualenv-clone==0.5.4
webencodings==0.5.1
xhtml2pdf==0.2.4
zope.interface==5.1.0
